Always use wheel-advertised catalog, remove dash folder config, misc dash UI cleanup
