+++ Devices/WheelUi/DashboardManagementControl.Files.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
@@ -88,6 +89,108 @@ namespace MozaPlugin.Devices.WheelUi
                     UploadLibraryCombo.SelectedIndex = 0;
             }
             _uploadLibrarySeeded = true;
+            UpdateUploadFolderInfo();
+        }
+
+        // ── Dashboard-library folder (relocated from the telemetry section) ──
+        // The mzdash folder is only needed to populate the upload library;
+        // telemetry binds to the wheel's live catalog, so the folder controls
+        // live here next to the library picker.
+
+        private void UploadSetFolder_Click(object sender, RoutedEventArgs e)
+        {
+            if (_plugin == null) return;
+            using (var dlg = new System.Windows.Forms.FolderBrowserDialog())
+            {
+                dlg.Description = "Select folder containing .mzdash dashboard files";
+                dlg.ShowNewFolderButton = false;
+                if (!string.IsNullOrEmpty(_plugin.ActiveTelemetryMzdashFolder)
+                    && System.IO.Directory.Exists(_plugin.ActiveTelemetryMzdashFolder))
+                    dlg.SelectedPath = _plugin.ActiveTelemetryMzdashFolder;
+                if (dlg.ShowDialog() != System.Windows.Forms.DialogResult.OK) return;
+                ApplyUploadFolder(dlg.SelectedPath);
+            }
+        }
+
+        private void UploadAutoDetect_Click(object sender, RoutedEventArgs e)
+        {
+            if (_plugin == null) return;
+
+            string dashesRoot = System.IO.Path.Combine(
+                Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
+                "MOZA Pit House", "_dashes");
+
+            if (!System.IO.Directory.Exists(dashesRoot))
+            {
+                MessageBox.Show(
+                    $"MOZA Pit House dashboard folder not found at:\n{dashesRoot}\n\n" +
+                    "Install MOZA Pit House and load a dashboard, or use Set Folder… to point at a custom location.",
+                    "Auto-detect dashboard folder",
+                    MessageBoxButton.OK, MessageBoxImage.Warning);
+                return;
+            }
+
+            byte[] uid = _plugin.Data?.WheelMcuUid ?? Array.Empty<byte>();
+            string uidHex = uid.Length == 12 ? WheelUiHelpers.UidToHex(uid) : "";
+
+            string? picked = null;
+            string? failReason = null;
+
+            if (!string.IsNullOrEmpty(uidHex))
+            {
+                string? match = System.IO.Directory.EnumerateDirectories(dashesRoot)
+                    .FirstOrDefault(p => string.Equals(
+                        new System.IO.DirectoryInfo(p).Name,
+                        uidHex,
+                        StringComparison.OrdinalIgnoreCase));
+                if (match != null)
+                    picked = match;
+                else
+                    failReason = $"No dashboard folder for the connected wheel (UID {uidHex}).\n" +
+                                 $"Looked under:\n{dashesRoot}\n\n" +
+                                 "Open a dashboard in MOZA Pit House for this wheel first.";
+            }
+            else
+            {
+                var guidDirs = System.IO.Directory.GetDirectories(dashesRoot)
+                    .Where(p => System.Text.RegularExpressions.Regex.IsMatch(
+                        new System.IO.DirectoryInfo(p).Name, "^[0-9a-fA-F]{24}$"))
+                    .ToList();
+                if (guidDirs.Count == 1)
+                    picked = guidDirs[0];
+                else if (guidDirs.Count == 0)
+                    failReason = "No wheel-specific dashboard folders found under _dashes. " +
+                                 "Open MOZA Pit House and load a dashboard, then try again.";
+                else
+                    failReason = $"Multiple dashboard folders found ({guidDirs.Count}) and no wheel is connected. " +
+                                 "Connect your wheel and try again, or use Set Folder… to choose manually.";
+            }
+
+            if (picked == null)
+            {
+                MessageBox.Show(failReason, "Auto-detect dashboard folder",
+                    MessageBoxButton.OK, MessageBoxImage.Information);
+                return;
+            }
+
+            ApplyUploadFolder(picked);
+        }
+
+        private void ApplyUploadFolder(string path)
+        {
+            if (_plugin == null) return;
+            _plugin.ActiveTelemetryMzdashFolder = path;
+            _plugin.SaveSettings();
+            _plugin.DashCache?.LoadFromFolder(path);
+            SeedUploadLibrary(force: true);
+            UpdateUploadFolderInfo();
+        }
+
+        private void UpdateUploadFolderInfo()
+        {
+            if (UploadFolderInfo == null) return;
+            var folder = _plugin?.ActiveTelemetryMzdashFolder;
+            UploadFolderInfo.Text = string.IsNullOrEmpty(folder) ? "" : $"Folder: {folder}";
         }
 
         private void UploadLibraryCombo_SelectionChanged(object sender, SelectionChangedEventArgs e)
